requirements.txt
playwright
pandas
python-dotenv
rich
mysql-connector-python
